Let features contribute their own nav entries
